add Eulerian_shared_classes for weakly-compressible and compressible fluid
