+++ tests/2d_examples/test_2d_eulerian_taylor_green/2d_eulerian_taylor_green.h
@@ -8,7 +8,8 @@
 #ifndef TAYLOR_GREEN_H
 #define TAYLOR_GREEN_H
 
-#include "sphinxsys.h"
+#include "common_shared_eulerian_classes.h" // basic eulerian classes for fluid.
+#include "common_compressible_eulerian_classes.h" // eulerian classes for compressible fluid only.
 using namespace SPH;
 //----------------------------------------------------------------------
 //	Basic geometry parameters and numerical setup.
